Moved Trace and TraceContextHolder to Heimdall-Core
